Add REST Client implementation
Step 1: we refactor Transport part

Closes #113.
